change magic number to 1 byte

+++ src/observer/sql/expr/tuple.h
@@ -211,7 +211,7 @@ public:
 
     // check if cell is null
     if (field_meta->nullable()) {
-      int32_t *null_flag = (int32_t *)(this->record_->data() + field_meta->offset());
+      uint8_t *null_flag = (uint8_t *)(this->record_->data() + field_meta->offset());
       if (*null_flag == NULL_MAGIC_NUMBER) {
         cell.set_null();
       }
